requirements.txt
torch>=1.8.0
numpy
pyyaml
matplotlib
sacrebleu
tqdm
